[IMPROVE] reactive mongo collector should update used size

+++ apache-james/extensions/mailets/src/main/java/com/disposableemail/apache/james/mailet/collector/pojo/Account.java
@@ -14,9 +14,8 @@ public class Account {
     private String mailboxId;
     private Boolean isDisabled;
     private Boolean isDeleted;
-    private Integer quota;
-    private Integer used;
+    private long quota;
+    private long used;
     private Instant createdAt;
     private Instant updatedAt;
-
 }
